app.R
The process of creating an RShiny application for sales prediction involves utilizing various elements in the R language. It begins with writing a script containing pieces of code to define the user interface (UI) and server logic, and then using R functions to process data and make predictions. Within this script, users will employ built-in R functions as well as possibly functions previously created to handle responses to user input and generate corresponding outputs. Follow the steps outlined in the syntax command to create a suitable RShiny application.